Move Badwords to new file
